Delivery 3 - Thursday 14:00

+++ apps/mail/views/MailIndex.jsx
@@ -17,12 +17,16 @@ export function MailIndex() {
     const [searchParams, setSearchParams] = useSearchParams()
     const [filterBy, setFilterBy] = useState(emailService.getFilterFromQueryString(searchParams))
     const [isFormVisible, setFormVisibility] = useState(false)
+    const [emailsFromStorage, setEmailsToStorage] = useState(null)
 
-    const { txt , status } = filterBy
+    const { txt, status } = filterBy
 
     useEffect(() => {
         loadEmails()
         setSearchParams(filterBy)
+        loadEmailsFromStorage()
+        // console.log('emails', emails)
+        // loadEmailsFromStorage()
 
         return () => {
             console.log('Bye Bye')
@@ -31,35 +35,61 @@ export function MailIndex() {
 
     function loadEmails() {
         emailService
-            .query(filterBy)
+            .queryFilterBy(filterBy)
             .then(emails => setEmails(emails))
             .catch(err => console.log('err:', err))
     }
-    
-    function onRemoveEmail(emailId) {
-        emailService.remove(emailId)
-        .then(() => {
-            setEmails(prevEmails => {
-                return prevEmails.filter(email => email.id !== emailId)
+
+    function loadEmailsFromStorage() {
+        emailService.query()
+            .then(emailsSotrage => {
+                setEmailsToStorage(emailsSotrage)
             })
-            showSuccessMsg(`Email successfully removed! ${emailId}`)
-        })
-        .catch(err => {
-            console.log('err:', err)
-            showErrorMsg(`Error while trying to delete an email ${emailId}`)
-            
-        })
     }
-    
+
+
+    function onRemoveEmail(emailId) {
+        // console.log('emails from storage', emailsFromStorage)
+        // console.log('emailId', emailId)
+        emailService.get(emailId)
+            .then((email) => {
+                console.log(email)
+                if (!email.removedAt) {
+                    email.removedAt = Math.floor(Date.now() / 1000)
+                    console.log(email)
+                    emailService.put(email)
+                        .then((email) => {
+                            console.log('email', email)
+                            showSuccessMsg(`Email moved to trash ${emailId}`)
+                            loadEmails()
+                        })
+                } else {
+                    emailService.remove(emailId)
+                        .then(() => {
+                            console.log('emailId', emails.removedAt)
+                            setEmails(prevEmails => {
+                                return prevEmails.filter(email => email.id !== emailId)
+                            })
+                            showSuccessMsg(`Email successfully removed! ${emailId}`)
+                        })
+                        .catch(err => {
+                            console.log('err:', err)
+                            showErrorMsg(`Error while trying to delete an email ${emailId}`)
+                        })
+                }
+            })
+    }
+
+
     const handleEmailComposeClick = () => {
         setFormVisibility(prevVisibility => !prevVisibility)
     }
-    
+
     function onSetFilter(filterBy) {
         setFilterBy(prevFilter => ({ ...prevFilter, ...filterBy }))
     }
 
-    console.log('filterBy', filterBy)
+    // console.log('filterBy', filterBy)
 
     if (!emails) return <div>Loading...</div>
     return (
